fix(config): ConfigBuilder.autoConfig is considered when instantiating new Config
`Config`'s autoConfigure field should behave as per expectations. We
need to make changes to ConfigBuilder and ConfigFluent in order to
achieve this.
- Add a new Constructor with additional boolean arguments `autoConfigure` and
  `shouldSetDefaultValues`
- Add a new class `SundrioConfig` that would extend `Config` for
  generating the Sundrio Builder for Config
- Move generated classes `ConfigBuilder` and `ConfigFluent` to `src/main/java`
  - `ConfigBuilder` would extend generated SundrioConfigBuilder and
    override the `build()` method by calling `disableAutoConfig()` if
    `autoConfigure` is nto provided by the user. Earlier `ConfigBuidler` was calling `new Config()`
    (this was enabling auto configuration)
- Make all fields in Config class use boxed types instead of primitive
  types so that we can distinguish whether they have been configured by
  user.


(cherry picked from commit 08b0e9f301fb163d9b98f71cd74fdf132ae7625b)

+++ openshift-client/src/main/java/io/fabric8/openshift/client/osgi/ManagedOpenShiftClient.java
@@ -131,7 +131,7 @@ public class ManagedOpenShiftClient extends NamespacedOpenShiftClientAdapter {
       builder.withOpenShiftUrl(URLUtils.join(builder.getMasterUrl(), "oapi", builder.getOapiVersion()));
     }
     if (properties.containsKey(OPENSHIFT_BUILD_TIMEOUT_SYSTEM_PROPERTY)) {
-      builder.withBuildTimeout(Integer.parseInt((String) properties.get(OPENSHIFT_BUILD_TIMEOUT_SYSTEM_PROPERTY)));
+      builder.withBuildTimeout(Long.parseLong((String) properties.get(OPENSHIFT_BUILD_TIMEOUT_SYSTEM_PROPERTY)));
     } else {
       builder.withBuildTimeout(DEFAULT_BUILD_TIMEOUT);
     }
